==> data_prcocess/00-gb-cnt-fix-t.ipynb <==
import numpy as np
import pandas as pd
from tqdm import tqdm
from pathlib import Path
import time
import os

df = pd.read_csv('./rawData/Foursquare_global/dataset_TIST2015_Checkins.txt', sep='\t', names=['uid', 'pid', 't', 'offset'])
df.info()

df['t_len'] = df.t.apply(lambda x: len(x))
f_df = df[df['t_len']==30]
len(df), len(f_df)

def tzo_convert(tzo_to_convert):
    return  datetime.timedelta(hours=tzo_to_convert/60)
f_df['check-in_time'] = f_df['t'].parallel_apply(lambda x: datetime.datetime.strptime(x, "%a %b %d %H:%M:%S +0000 %Y"))
f_df['check-in_time'] = f_df.parallel_apply(lambda x: str(x['check-in_time'] + tzo_convert(x['offset'])), axis=1)
f_df['day'] = f_df['check-in_time'].parallel_apply(lambda x : x[:10])
f_df = f_df.rename(columns = {'uid' : 'user', 'pid' : 'location_id'})
f_df.to_csv('./rawData/Foursquare_global/check_in.csv')
f_df.head()

df = pd.read_csv('./rawData/Foursquare_global/check_in.csv')
df.head()

p_df = pd.read_csv("./rawData/Foursquare_global/dataset_TIST2015_POIs.txt", sep='\t', names=['location_id', 'p_lat', 'p_lon', 'p_type', 'countrycode'])
check_gps_df = pd.merge(df, p_df, on = 'location_id')
print(len(df), len(p_df), len(check_gps_df))
check_gps_df.head()

c_df = pd.read_csv("./rawData/Foursquare_global/dataset_TIST2015_Cities.txt", sep='\t', names=['city', 'c_lat', 'c_lon', 'c_countrycode', 'c_country', 'c_type'])
city_tuple = list(tuple(zip(c_df['c_lat'], c_df['c_lon'], c_df['city'])))
c_df.head()

import math
def get_city(lat, lon, city_tup):
    min_dist = float('inf')
    for city in city_tup:
        dist = math.sqrt((lat - city[0])**2 + (lon - city[1])**2)
        if dist < min_dist:
            min_dist = dist
            loc_city = city[2]
    return loc_city
check_gps_df['city'] = check_gps_df.parallel_apply(lambda x : get_city(x['p_lat'], x['p_lon'], city_tuple), axis=1)
res_df = pd.merge(check_gps_df, c_df, on = 'city')
res_df.head()

res_df = res_df.rename(columns = {'countrycode' : 'p_countrycode'})
res_df.to_csv(f'./rawData/Foursquare_global/check_in_merge.csv', index=False)

dir = f'./rawData/Foursquare_global'
sub_dir = 'city_user_day_fix'
Path(f'{dir}/{sub_dir}/res').mkdir(parents=True, exist_ok=True)

df = pd.read_csv(f'{dir}/check_in_merge.csv')
part_df = df[['user', 'day', 'check-in_time']].reset_index().rename(columns={'index':'ori_id'})
part_df['check_in_time'] = pd.to_datetime(part_df['check-in_time'])
part_df['day'] = pd.to_datetime(part_df['day'])
part_df = part_df.sort_values(['user', 'day', 'check-in_time'])
part_df['hour'] = part_df['check-in_time'].apply(lambda x: x.hour)
part_df['minute'] = part_df['check-in_time'].apply(lambda x: x.minute)
hour_mask = part_df['hour'].apply(lambda x: x < 23).astype(int)
hour_offset = part_df['minute'].apply(lambda x: 1 if x > 30 else 0)
minute_offset = part_df['minute'].apply(lambda x: 60 - x if x > 30 else x)
part_df['hour_within_30min'] = part_df['hour'] + hour_offset * hour_mask
part_df['minute_offset'] = minute_offset
part_df[['hour', 'minute', 'hour_within_30min', 'minute_offset']].describe()

part_df = part_df.sort_values(['user', 'day', 'hour_within_30min', 'minute_offset'])
hour_df = part_df.drop_duplicates(['user', 'day', 'hour_within_30min'], keep='first').reset_index(drop=True)
hour_df.info()

gb_cnt_sorted = hour_df.groupby(['user', 'day']).count().sort_values(['user', 'day']).reset_index()
gb_cnt_sorted = gb_cnt_sorted.rename(columns={'check-in_time':'user_day_count'})
gb_cnt_sorted[['user', 'day', 'user_day_count']].to_csv(f'{dir}/{sub_dir}/res/gb_cnt_sorted.csv', index=False)
ans_df = df.iloc[hour_df['ori_id'].to_numpy()]
ans_df.info()

ans_df['hour'] = hour_df['hour_within_30min']
ans_df['hour_offset'] = hour_df['minute_offset']

ans_df.to_csv(f'{dir}/{sub_dir}/res/gb_sorted.csv', index=False)
