==> data_prcocess/01-city-csv-fix-t.ipynb <==
import numpy as np
import pandas as pd
from tqdm import tqdm
from pathlib import Path
import time
import os
import datetime


dir = '../rawData/Foursquare_global'
sub_dir = 'city_user_day_fix'
df = pd.read_csv(f'{dir}/{sub_dir}/gb_sorted.csv') 
df.info()

df.head(3)

df['day'] = pd.to_datetime(df['day'])
df['check-in_time'] = pd.to_datetime(df['check-in_time'])
df = df.sort_values(['user', 'day', 'check-in_time']).reset_index(drop=True)
df['ori_id'] = np.arange(len(df))
ori_id = df['ori_id']

grouped = df.groupby(['user', 'day'])
gb_cnt_sorted = grouped.count().reset_index()

group_ori_id = np.concatenate([g['ori_id'].values for name, g in grouped])
np.all(group_ori_id == df['ori_id'].values)

gb_cnt_sorted['user_day_cnt'] = gb_cnt_sorted['check-in_time']
traj_id = np.concatenate([np.full((cnt,), i) for i, cnt in enumerate(gb_cnt_sorted['user_day_cnt'].tolist())])
df['traj_id'] = traj_id
len(df), len(ori_id), len(traj_id)

c_df = pd.read_csv("./rawData/Foursquare_global/dataset_TIST2015_Cities.txt", sep='\t', names=['city', 'c_lat', 'c_lon', 'c_countrycode', 'c_country', 'c_type'])
city_records = dict()
for city in c_df['city'].tolist():
    city_records[city] = []
    
df.info()

for city in tqdm(c_df['city'].tolist()):
    city_df = df[df['city'] == city]
    city_traj_id = set(city_df['traj_id'].unique().tolist())
    city_traj_df = df[df['traj_id'].isin(city_traj_id)]
    city_traj_df.to_csv(f'{dir}/{sub_dir}/res/{city}.csv', index=False)
