==> data_prcocess/02-select-data.ipynb <==
import numpy as np
import pandas as pd
from tqdm import tqdm
from pathlib import Path
import time
import os
import json

dir = './rawData/Foursquare_global'
sub_dir = 'city_user_day_fix'
path = f'{dir}/{sub_dir}/res'
os.chdir(path)
num = 0
for f in sorted(os.listdir()):
    if f.endswith('.csv'):
        df = pd.read_csv(f)
        cnt = np.unique(df['traj_id'], return_counts=True)
        len_larger_4 = (cnt[1]>=4).sum()
        len_larger_6 = (cnt[1]>=6).sum()
        len_larger_8 = (cnt[1]>=8).sum()
        if len_larger_8 < 1000 and f[:-4] != 'New York':
            continue
        print(f'{f:>20} {len(df):08d} {len_larger_4:08d} {len_larger_6:08d} {len_larger_8:08d}')
        num += 1
print(num)

need_cities = ['Bangkok', 'Istanbul', 'Jakarta', 'Kuala Lumpur', 'Moscow', 'New York', 'Osaka', 'Shah Alam', 'Singapore', 'Tokyo']
c_df = pd.read_csv("./rawData/Foursquare_global/dataset_TIST2015_Cities.txt", sep='\t', names=['city', 'c_lat', 'c_lon', 'c_countrycode', 'c_country', 'c_type'])
df = c_df[c_df['city'].isin(need_cities)]
df

dir = './rawData/Foursquare_global/city_user_day_fix/res/'
for f in tqdm(sorted(os.listdir(dir))):
    if f.endswith('csv'):
        new_f = f.replace(" ", "_")
        os.rename(os.path.join(dir, f), os.path.join(dir, new_f))

TEST_CITY = ['New York']
TEST_CITY = [c.replace(" ", "_") for c in TEST_CITY]
TEST_CITY

data = 'New York'
city = data.replace(" ", "_")
city

df = pd.read_csv(f'./rawData/Foursquare_global/city_user_day_fix/res/{city}.csv')
df.head()

# val, cnt = np.unique(df['traj_id'].to_numpy(), return_counts=True)
traj_id_f = np.unique(cnt[0][np.where(cnt[1]>=6)[0]])
df_f = df[df['traj_id'].isin(traj_id_f)]
df_f.head()

df_f = df_f.sort_values(by = ['user', 'day', 'check-in_time']).reset_index(drop=True)
pois = np.unique(df_f['location_id'])
pois_id_dict = dict(zip(pois, range(len(pois))))
df_f['poi_id'] = df_f['location_id'].apply(lambda x:pois_id_dict[x])
Path('./rawData/Foursquare_global/city_user_day_fix/cities/').mkdir(parents=True, exist_ok=True)
df_f.to_csv(f'./rawData/Foursquare_global/city_user_day_fix/cities/{city}_f.csv', index=False)
grouped = df_f.groupby(['user', 'day'])
u_l, seq_l, t_l = [], [], []
for name, g in grouped:
    u_l.append(g['user'].values.tolist()[0])
    seq_l.append(g['poi_id'].values.tolist())
    t_l.append(g['check-in_time'].values.tolist())
len(u_l), len(seq_l), len(seq_l)

m_tes_tra_ratio, m_tes_val_ratio, m_tes_tes_ratio = 0.7, 0.1, 0.2
all_idx = [i for i in range(len(seq_l))]
np.random.shuffle(all_idx)
if city in TEST_CITY:
    tra_idx = all_idx[ : int(len(seq_l) * m_tes_tra_ratio)]
    tra_user = [u_l[id] for id in tra_idx]
    tra_path = [seq_l[id] for id in tra_idx]
    tra_tim = [t_l[id] for id in tra_idx]
    
    val_idx = all_idx[int(len(seq_l) * m_tes_tra_ratio) : int(len(seq_l) * (m_tes_tra_ratio + m_tes_val_ratio))]
    val_user = [u_l[id] for id in val_idx]
    val_path = [seq_l[id] for id in val_idx] 
    val_tim = [t_l[id] for id in val_idx]

    tes_idx = all_idx[int(len(seq_l) * (m_tes_tra_ratio + m_tes_val_ratio)):] 
    tes_user = [u_l[id] for id in tes_idx]
    tes_path = [seq_l[id] for id in tes_idx]
    tes_tim = [t_l[id] for id in tes_idx]
    
data = {'tra_user': tra_user, 'tra_path': tra_path, 'tra_tim': tra_tim, 
        'val_user': val_user, 'val_path': val_path, 'val_tim': val_tim,
        'tes_user': tes_user, 'tes_path': tes_user, 'tes_tim': tes_user}
data_str = json.dumps(data)
with open(f'./rawData/Foursquare_global/city_user_day_fix/cities/{city}_split.json', 'w') as file:
    file.write(data_str)
